requirements.txt
Django==5.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
pymysql>=1.1.0
mysqlclient>=2.2.0
Pillow==10.1.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pandas==2.1.3
openpyxl==3.1.2
zipfile36==0.1.3; python_version < '3.0'